zng_wgt_access

Function on_access_expander

pub fn on_access_expander(
    child: impl IntoUiNode,
    handler: impl WidgetHandler<AccessExpanderArgs>,
) -> UiNode

P Access requested to expand or collapse the widget content.

Preview

You can preview this event using on_pre_access_expander. Otherwise the handler is only called after the widget content has a chance to stop propagation.

Async

You can use async event handlers with this property.