zng_wgt_access

Function scroll_vertical

pub fn scroll_vertical(
    child: impl UiNode,
    normal_y: impl IntoVar<Factor>,
) -> impl UiNode

P Sets the amount scrolled vertically if allowed.

The normal_y value can be a read-only variable, the variable can be updated without needing to rebuild info for every pixel scrolled, if the view-process requires access info the value is updated every render together with the widget bounds updates.

The value must be normalized in the 0..=1 range, 0 is showing the content topmost edge, 1 is showing the content the bottommost edge.