Crate zng_wgt_access

Properties that define accessibility metadata.

The properties in this crate should only be used by widget implementers, they only define metadata for accessibility, this metadata signals the availability of behaviors that are not implemented by these properties, for example an AccessRole::Button widget must also be focusable and handle click events, an AccessRole::TabList must contain widgets marked AccessRole::Tab.

Crate

This crate is part of the zng project.

Structs

AutoComplete

Defines how inputting text could trigger display of one or more predictions of the user’s intended value.

Invalid

Defines the kind of invalid data error of a widget.

Enums

AccessCmdName

Accessibility command without associated data.

AccessRole

Accessibility role of a node in the accessibility tree.

CurrentKind

Kind of current item a widget represents.

LiveIndicator

Defines how a live update is communicated to the user.

Orientation

Widget orientation.

Popup

Popup type.

SortDirection

Sort direction.

Functions

access_commands

P Append supported access commands.

access_role

P Sets the widget kind for accessibility services.

accessible

P Defines if the widget and descendants can be present in the accessibility info tree.

active_descendant

P Identifies the currently active widget when focus is on a composite widget.

auto_complete

P Set how input text triggers display of one or more predictions of the user’s intended value for a ComboBox, SearchBox, or TextInput.

checked

P If the widget is checked (Some(true)), unchecked (Some(false)), or if the checked status is indeterminate (None).

col_count

P Sets the total number of columns in a Table, Grid, or TreeGrid when not all columns are present in tree.

col_index

P Sets the widget’s column index in the parent table or grid.

col_span

P Sets the number of columns spanned by the widget in the parent table or grid.

controls

P Append widgets whose contents or presence are controlled by this widget to the controlled list.

current

P Indicates that the widget represents the current item of a kind.

described_by

P Append widgets that describes this widget to the descriptors list.

details

P Append widgets that provide additional information related to this widget to the details list.

error_message

P Indicates that the widget is an error message for the invalid_wgt.

expanded

P Indicate that the widget toggles the visibility of related widgets.

flows_to

P Append options for next widget to be read by screen readers.

invalid

P Indicates that the widget’s data is invalid with optional kinds of errors.

item_count

P Sets the number of items in the current set of list items or tree items when not all items in the set are present in the tree.

item_index

P Sets the widget’s number or position in the current set of list items or tree items when not all items are present in the tree.

label

P Sets a custom name for the widget in accessibility info.

labelled_by

P Append widgets that provide additional information related to this widget.

labelled_by_child

P Uses the accessible children as labelled_by.

lang

P Defines the language used by screen-readers to read text in this widget and descendants.

level

P Sets the hierarchical level of the widget within a parent scope.

live

P Indicate that the widget can change, how the change can be announced, if atomic the entire widget must be re-read, if busy the screen reader must wait until the change completes.

modal

P Sets if the widget is modal when displayed.

multi_selectable

P Indicates that the user may select more than one item from the current selectable descendants.

on_access_click

P Access requested a click.

on_access_expander

P Access requested to expand or collapse the widget content.

on_access_increment

P Access requested to increment or decrement the widget value by steps.

on_access_number

P Access requested a number input.

on_access_scroll

P Access requested a scroll command.

on_access_selection

P Access requested a text selection.

on_access_text

P Access requested a text input/replace.

on_access_tooltip

P Access requested to show or hide the widget’s tooltip.

on_pre_access_click

P Preview on_access_click event.

on_pre_access_expander

P Preview on_access_expander event.

on_pre_access_increment

P Preview on_access_increment event.

on_pre_access_number

P Preview on_access_number event.

on_pre_access_scroll

P Preview on_access_scroll event.

on_pre_access_selection

P Preview on_access_selection event.

on_pre_access_text

P Preview on_access_text event.

on_pre_access_tooltip

P Preview on_access_tooltip event.

orientation

P Indicates whether the widget’s orientation is horizontal, vertical, or unknown/ambiguous.

owns

P Append owned widgets that are children of this widget, but are not already children in the info tree.

placeholder

P Short hint (a word or short phrase) intended to help the user with data entry when a form control has no value.

popup

P Indicates the availability and type of interactive popup widget.

read_only

P Indicates that the widget is not editable, but is otherwise operable.

required

P Indicates that user input is required on the widget before a form may be submitted.

row_count

P Sets the total number of rows in a Table, Grid, or TreeGrid when not all rows are present in the tree.

row_index

P Sets the widget’s row index in the parent table or grid.

row_span

P Sets the number of rows spanned by the widget in the parent table or grid.

scroll_horizontal

P Sets the amount scrolled horizontally if allowed.

scroll_vertical

P Sets the amount scrolled vertically if allowed.

selected

P Indicates that the widget is selected.

sort

P Sets the sort direction for the table or grid items.

value

P Set the current value.

value_max

P Set the maximum value (inclusive).

value_min

P Set the minimum value (inclusive).

value_text

P Set a text that is a readable version of the current value.